ingestion service refactored, tests
